+++ tests/testthat/test-format_parameters.R
@@ -268,4 +268,59 @@ if (requiet("testthat") && requiet("parameters") && requiet("splines")) {
       SpeciesNo_Specieses = "Species [No_Specieses]", Petal.Width = "Petal Width"
     ))
   })
+
+
+  if (requiet("datawizard")) {
+    test_that("format_parameters-labelled data-1", {
+      data(efc, package = "datawizard")
+      m <- lm(neg_c_7 ~ e42dep + c172code, data = efc)
+      mp <- model_parameters(m)
+
+      out <- utils::capture.output(print(mp, pretty_names = "labels"))
+      expect_equal(
+        trimws(unlist(strsplit(out[4], "|", fixed = TRUE))[1]),
+        "elder's dependency [slightly dependent]"
+      )
+      expect_equal(
+        trimws(unlist(strsplit(out[5], "|", fixed = TRUE))[1]),
+        "elder's dependency [moderately dependent]"
+      )
+
+      out <- utils::capture.output(print(mp))
+      expect_equal(
+        trimws(unlist(strsplit(out[4], "|", fixed = TRUE))[1]),
+        "e42dep [2]"
+      )
+      expect_equal(
+        trimws(unlist(strsplit(out[5], "|", fixed = TRUE))[1]),
+        "e42dep [3]"
+      )
+    })
+
+    test_that("format_parameters-labelled data-2", {
+      data(iris)
+      m <- lm(Sepal.Width ~ Species + Sepal.Length, data = iris)
+      mp <- model_parameters(m)
+
+      out <- utils::capture.output(print(mp, pretty_names = "labels"))
+      expect_equal(
+        trimws(unlist(strsplit(out[4], "|", fixed = TRUE))[1]),
+        "Species [versicolor]"
+      )
+      expect_equal(
+        trimws(unlist(strsplit(out[5], "|", fixed = TRUE))[1]),
+        "Species [virginica]"
+      )
+
+      out <- utils::capture.output(print(mp))
+      expect_equal(
+        trimws(unlist(strsplit(out[4], "|", fixed = TRUE))[1]),
+        "Species [versicolor]"
+      )
+      expect_equal(
+        trimws(unlist(strsplit(out[5], "|", fixed = TRUE))[1]),
+        "Species [virginica]"
+      )
+    })
+  }
 }